+++ pkg/providers/property/testdata/TestResDOMLateValidation/import.tf
@@ -1,0 +1,6 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+resource "akamai_property_domainownership_late_validation" "test" {
+}